deneme2
depositorsla ilgili herşey(dto/model/controller/repository güncellendi(hatalı ola bilir)).depository index html sayfasi ve update sayfasi eklendi(update bozuk)

+++ sbweb/src/main/java/org/sau/sbweb/repositories/DepositorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepositorRepository extends JpaRepository<Depositor, Integer> {
+public interface DepositorRepository extends JpaRepository<Depositor, Long> {
 }
